+++ src/test/java/lotto/domainTest/LottoTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.domainTest;
 
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
